+++ src/main/java/com/refinedmods/refinedstorage/mekanism/grid/ChemicalGridResource.java
@@ -20,7 +20,6 @@ import com.refinedmods.refinedstorage.mekanism.ChemicalResourceType;
 import com.refinedmods.refinedstorage.mekanism.ChemicalUtil;
 
 import java.util.List;
-import java.util.Map;
 import java.util.Optional;
 import java.util.Set;
 import java.util.function.Function;